Доработал код исходя из замечаний
