+++ precode_test.go
@@ -1,49 +1,21 @@
 package main
 
 import (
+	"github.com/stretchr/testify/require"
 	"net/http"
 	"net/http/httptest"
-	"strconv"
 	"strings"
 	"testing"
-
-	"github.com/stretchr/testify/assert"
 )
 
-var cafeList = map[string][]string{
-	"moscow": {"Мир кофе", "Сладкоежка", "Кофе и завтраки", "Сытый студент"},
-}
+func TestMainHandlerWhenSuccess(t *testing.T) {
+	req := httptest.NewRequest("GET", "/cafe?count=10&city=moscow", nil)
+	responseRecorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(mainHandle)
+	handler.ServeHTTP(responseRecorder, req)
 
-func mainHandle(w http.ResponseWriter, req *http.Request) {
-	countStr := req.URL.Query().Get("count")
-	if countStr == "" {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("count missing"))
-		return
-	}
-
-	count, err := strconv.Atoi(countStr)
-	if err != nil {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("wrong count value"))
-		return
-	}
-
-	city := req.URL.Query().Get("city")
-	cafe, ok := cafeList[city]
-	if !ok {
-		w.WriteHeader(http.StatusBadRequest)
-		w.Write([]byte("wrong city value"))
-		return
-	}
-
-	if count > len(cafe) {
-		count = len(cafe)
-	}
-
-	answer := strings.Join(cafe[:count], ",")
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte(answer))
+	require.Equal(t, http.StatusOK, responseRecorder.Code)
+	require.NotEmpty(t, responseRecorder.Body.String())
 }
 
 func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
@@ -52,9 +24,9 @@ func TestMainHandlerWhenCountMoreThanTotal(t *testing.T) {
 	handler := http.HandlerFunc(mainHandle)
 	handler.ServeHTTP(responseRecorder, req)
 
-	assert.Equal(t, http.StatusOK, responseRecorder.Code)
-	assert.NotEmpty(t, responseRecorder.Body.String())
-	assert.Len(t, strings.Split(responseRecorder.Body.String(), ","), 4) // Проверяем, что вернулось 4 кафе
+	require.Equal(t, http.StatusOK, responseRecorder.Code)
+	require.NotEmpty(t, responseRecorder.Body.String())
+	require.Len(t, strings.Split(responseRecorder.Body.String(), ","), 4) // Проверяем, что вернулось 4 кафе
 }
 
 func TestMainHandlerWhenCityNotSupported(t *testing.T) {
@@ -63,8 +35,8 @@ func TestMainHandlerWhenCityNotSupported(t *testing.T) {
 	handler := http.HandlerFunc(mainHandle)
 	handler.ServeHTTP(responseRecorder, req)
 
-	assert.Equal(t, http.StatusBadRequest, responseRecorder.Code)
-	assert.Equal(t, "wrong city value", responseRecorder.Body.String())
+	require.Equal(t, http.StatusBadRequest, responseRecorder.Code)
+	require.Equal(t, "wrong city value", responseRecorder.Body.String())
 }
 
 func TestMainHandlerWhenCountMissing(t *testing.T) {
@@ -73,8 +45,8 @@ func TestMainHandlerWhenCountMissing(t *testing.T) {
 	handler := http.HandlerFunc(mainHandle)
 	handler.ServeHTTP(responseRecorder, req)
 
-	assert.Equal(t, http.StatusBadRequest, responseRecorder.Code)
-	assert.Equal(t, "count missing", responseRecorder.Body.String())
+	require.Equal(t, http.StatusBadRequest, responseRecorder.Code)
+	require.Equal(t, "count missing", responseRecorder.Body.String())
 }
 
 func TestMainHandlerWhenCountWrongValue(t *testing.T) {
@@ -83,6 +55,6 @@ func TestMainHandlerWhenCountWrongValue(t *testing.T) {
 	handler := http.HandlerFunc(mainHandle)
 	handler.ServeHTTP(responseRecorder, req)
 
-	assert.Equal(t, http.StatusBadRequest, responseRecorder.Code)
-	assert.Equal(t, "wrong count value", responseRecorder.Body.String())
+	require.Equal(t, http.StatusBadRequest, responseRecorder.Code)
+	require.Equal(t, "wrong count value", responseRecorder.Body.String())
 }
